# scalar quantization: fix dot byte implementation