--- a/adapters/repos/db/vector/hnsw/distancer/c/dot_byte_avx256.c
+++ b/adapters/repos/db/vector/hnsw/distancer/c/dot_byte_avx256.c
@@ -22,7 +22,7 @@ void dot_byte_256(unsigned char *a, unsigned char *b, unsigned int *res, long *l
         long acc = 0;
         for (int i = 0; i < n; i++)
         {
-            acc += a[i] * b[i];
+            acc += (unsigned int)(a[i]) * (unsigned int)(b[i]);
         }
 
         *res = acc;
@@ -42,14 +42,14 @@ void dot_byte_256(unsigned char *a, unsigned char *b, unsigned int *res, long *l
         vec_b = _mm256_loadu_si256((const __m256i *)(b + i));
 
         // Create two registries for vector a
-        __m256i a_high = _mm256_srai_epi16(vec_a, 8);  // arithmetic right shift sign extends
+        __m256i a_high = _mm256_srli_epi16(vec_a, 8);  // arithmetic right shift
         __m256i a_low = _mm256_bslli_epi128(vec_a, 1); // left 1 byte = low to high in each 16-bit element
-        a_low = _mm256_srai_epi16(a_low, 8);           // arithmetic right shift sign extends
+        a_low = _mm256_srli_epi16(a_low, 8);           // arithmetic right shift
 
         // Create two registries for vector b
-        __m256i b_high = _mm256_srai_epi16(vec_b, 8);
+        __m256i b_high = _mm256_srli_epi16(vec_b, 8);
         __m256i b_low = _mm256_bslli_epi128(vec_b, 1);
-        b_low = _mm256_srai_epi16(b_low, 8);
+        b_low = _mm256_srli_epi16(b_low, 8);
 
         __m256i prod_hi = _mm256_madd_epi16(a_high, b_high);
         __m256i prod_lo = _mm256_madd_epi16(a_low, b_low);
@@ -66,12 +66,12 @@ void dot_byte_256(unsigned char *a, unsigned char *b, unsigned int *res, long *l
     acc128 = _mm_add_epi32(acc128, _mm_shuffle_epi32(acc128, _MM_SHUFFLE(0, 1, 2, 3)));
     acc128 = _mm_add_epi32(acc128, _mm_shuffle_epi32(acc128, _MM_SHUFFLE(0, 0, 0, 1)));
 
-    int result = _mm_extract_epi32(acc128, 0);
+    unsigned int result = _mm_extract_epi32(acc128, 0);
 
     // Tail
     for (; i < n; i++)
     {
-        result += a[i] * b[i];
+        result += (unsigned int)(a[i]) * (unsigned int)(b[i]);
     }
 
     *res = result;