… adapters/repos/db/vector/hnsw/distancer/asm/dot_byte_avx256.s