### What's being changed:

The current implementation wasn't tested properly because of a problem in a test.

```
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Xeon(R) CPU @ 2.60GHz
                          │ sq_dot_go.txt │         sq_dot_avx256.txt          │
                          │    sec/op     │   sec/op     vs base               │
DotByte/2_dimensions-4        4.463n ± 0%   2.669n ± 1%  -40.20% (p=0.002 n=6)
DotByte/4_dimensions-4        5.044n ± 0%   3.021n ± 0%  -40.12% (p=0.002 n=6)
DotByte/6_dimensions-4        8.075n ± 2%   5.058n ± 0%  -37.36% (p=0.002 n=6)
DotByte/8_dimensions-4        7.423n ± 0%   6.043n ± 3%  -18.59% (p=0.002 n=6)
DotByte/10_dimensions-4       8.604n ± 0%   6.549n ± 0%  -23.89% (p=0.002 n=6)
DotByte/12_dimensions-4       9.796n ± 2%   7.317n ± 0%  -25.30% (p=0.002 n=6)
DotByte/16_dimensions-4       12.16n ± 0%   11.55n ± 0%   -4.98% (p=0.002 n=6)
DotByte/24_dimensions-4       16.92n ± 0%   15.84n ± 0%   -6.41% (p=0.002 n=6)
DotByte/30_dimensions-4       20.48n ± 0%   19.12n ± 0%   -6.64% (p=0.002 n=6)
DotByte/32_dimensions-4      21.650n ± 0%   6.365n ± 0%  -70.60% (p=0.002 n=6)
DotByte/128_dimensions-4     78.605n ± 0%   8.838n ± 0%  -88.76% (p=0.002 n=6)
DotByte/256_dimensions-4     160.55n ± 0%   14.24n ± 1%  -91.13% (p=0.002 n=6)
DotByte/300_dimensions-4     186.75n ± 0%   21.34n ± 0%  -88.57% (p=0.002 n=6)
DotByte/384_dimensions-4     236.85n ± 1%   18.98n ± 0%  -91.99% (p=0.002 n=6)
DotByte/512_dimensions-4     313.35n ± 0%   24.36n ± 1%  -92.22% (p=0.002 n=6)
DotByte/768_dimensions-4     466.25n ± 1%   33.42n ± 0%  -92.83% (p=0.002 n=6)
DotByte/1024_dimensions-4    616.05n ± 0%   42.93n ± 0%  -93.03% (p=0.002 n=6)
DotByte/1536_dimensions-4    920.60n ± 0%   61.83n ± 1%  -93.28% (p=0.002 n=6)
geomean                       44.56n        12.06n       -72.94%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.